refactor: remove legacy shared components
