--- frontend/src/shared/components/Header.tsx
@@ -1,82 +1,0 @@
-import React from 'react';
-import { Bell, Search, User, LogOut } from 'lucide-react';
-import { Link } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import { useAuth } from '../../features/auth/hooks';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '../../components/ui/dropdown-menu';
-import { Button } from '../../components/ui/button';
-import { Input } from '../../components/ui/input';
-import { LanguageToggle } from '@/components/ui/language-toggle';
-import { ThemeToggle } from '@/components/ui/theme-toggle';
-import { BrandLogo } from '@/components/common/BrandLogo';
-
-export const Header: React.FC = () => {
-  const { user, logout, isAuthenticated } = useAuth();
-  const { t } = useTranslation();
-
-  const handleLogout = async () => {
-    await logout();
-  };
-
-  return (
-    <header className="bg-background border-b border-border px-6 py-4">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-4 flex-1">
-          <BrandLogo variant="icon" className="h-8" />
-          {/* Search */}
-          <div className="flex-1 max-w-md">
-            <div className="relative">
-              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground" size={16} />
-              <Input
-                placeholder={t('common.search')}
-                className="pl-10"
-              />
-            </div>
-          </div>
-        </div>
-
-        {/* Right side */}
-        <div className="flex items-center space-x-4 rtl:space-x-reverse">
-          <LanguageToggle />
-          <ThemeToggle />
-          {isAuthenticated ? (
-            <>
-              <Button asChild size="sm" variant="secondary">
-                <Link to="/dashboard">{t('common.dashboard')}</Link>
-              </Button>
-              <Button variant="ghost" size="sm">
-                <Bell size={16} />
-              </Button>
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button variant="ghost" size="sm" className="flex items-center space-x-2 rtl:space-x-reverse">
-                    <User size={16} />
-                    {user && (
-                      <span className="hidden md:inline">
-                        {user.firstName} {user.lastName}
-                      </span>
-                    )}
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent align="end">
-                  <DropdownMenuItem>
-                    <User size={16} className="mr-2" />
-                    {t('common.profile')}
-                  </DropdownMenuItem>
-                  <DropdownMenuItem onClick={handleLogout}>
-                    <LogOut size={16} className="mr-2" />
-                    {t('common.logout')}
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </>
-          ) : (
-            <Button asChild size="sm">
-              <Link to="/login">{t('common.login')}</Link>
-            </Button>
-          )}
-        </div>
-      </div>
-    </header>
-  );
-};
