--- frontend/src/shared/components/Layout.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import { Sidebar } from './Sidebar';
-import { Header } from './Header';
-import { useSidebar } from '../hooks/useSidebar';
-
-interface LayoutProps {
-  children?: React.ReactNode;
-}
-
-const Layout: React.FC<LayoutProps> = ({ children }) => {
-  const { isOpen } = useSidebar();
-
-  return (
-    <div className="min-h-screen bg-background">
-      <Sidebar />
-      <div className={`transition-all duration-300 ${isOpen ? 'lg:ml-64' : 'lg:ml-16'}`}>
-        <Header />
-        <main className="p-6">
-          {children || <Outlet />}
-        </main>
-      </div>
-    </div>
-  );
-};
-
-export default Layout;
