--- frontend/src/shared/components/Spinner.tsx
@@ -1,31 +1,0 @@
-import React from 'react';
-import { Scale } from 'lucide-react';
-
-interface LoadingSpinnerProps {
-  size?: 'sm' | 'md' | 'lg';
-  text?: string;
-  className?: string;
-}
-
-export const LoadingSpinner: React.FC<LoadingSpinnerProps> = ({ 
-  size = 'md', 
-  text,
-  className = ''
-}) => {
-  const sizeClasses = {
-    sm: 'w-4 h-4',
-    md: 'w-8 h-8',
-    lg: 'w-12 h-12'
-  };
-
-  return (
-    <div className={`flex flex-col items-center justify-center gap-2 ${className}`}>
-      <Scale className={`${sizeClasses[size]} animate-spin text-primary`} />
-      {text && (
-        <p className="text-sm text-muted-foreground animate-pulse">
-          {text}
-        </p>
-      )}
-    </div>
-  );
-};
