--- frontend/src/shared/hooks/useSidebar.ts
@@ -1,23 +1,0 @@
-import { create } from 'zustand';
-import { persist } from 'zustand/middleware';
-
-interface SidebarState {
-  isOpen: boolean;
-  toggle: () => void;
-  open: () => void;
-  close: () => void;
-}
-
-export const useSidebar = create<SidebarState>()(
-  persist(
-    (set) => ({
-      isOpen: true,
-      toggle: () => set((state) => ({ isOpen: !state.isOpen })),
-      open: () => set({ isOpen: true }),
-      close: () => set({ isOpen: false })
-    }),
-    {
-      name: 'sidebar-storage'
-    }
-  )
-);
